Revert "Feature/refactor drawer bar"
