+++ src/components/GUI.js
@@ -1,15 +1,94 @@
 import React from 'react';
+import { makeStyles } from '@material-ui/core/styles';
+
 import Typography from '@material-ui/core/Typography';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import Container from '@material-ui/core/Container';
+
+import Drawer from '@material-ui/core/Drawer';
+import List from '@material-ui/core/List';
+import Divider from '@material-ui/core/Divider';
+import ListItem from '@material-ui/core/ListItem';
+import ListItemIcon from '@material-ui/core/ListItemIcon';
+import ListItemText from '@material-ui/core/ListItemText';
+import clsx from 'clsx';
+
+import HomeIcon from '@material-ui/icons/Home';
+import PublishIcon from '@material-ui/icons/Publish';
+import KeyboardIcon from '@material-ui/icons/Keyboard';
+import SettingsIcon from '@material-ui/icons/Settings';
+import InfoIcon from '@material-ui/icons/Info';
+import ExitToAppIcon from '@material-ui/icons/ExitToApp';
+
 import GuiHome from './home';
 import GUIhotkeys from './hotkeys';
+
 import { ThemeProvider } from '@material-ui/styles';
 
 import light from './light'
 import dark from './dark'
+
 import MenuBar from './MenuBar'
-import DrawerBar from './DrawerBar'
+
+function DrawerBar(props) {
+  const useStyles = makeStyles({
+    list: {
+      width: 250,
+    },
+    fullList: {
+      width: 'auto',
+    },
+  });
+  const classes = useStyles();
+
+  const toggleDrawer = (open) => (event) => {
+    props.onClick(open);
+  };
+
+  let menu1 = {
+    1: ['Soundboard',<HomeIcon />],
+    2: ['Hotkeys',<KeyboardIcon />],
+    3: ['Upload',<PublishIcon />],
+    4: ['Settings',<SettingsIcon />],
+ }
+
+  const list = () => (
+    <div 
+    className={clsx(classes.list)} 
+    role="presentation" 
+    onClick={toggleDrawer(false)}
+    onKeyDown={toggleDrawer(false)}
+    >
+      <List>
+      {Object.keys(menu1).map((text, index) => (
+          <ListItem button key={menu1[text][0]} onClick={() => props.handleMenuClick(menu1[text][0])}>
+            <ListItemIcon>{menu1[text][1]}</ListItemIcon>
+            <ListItemText primary={menu1[text][0]} />
+          </ListItem>
+        ))}
+      </List>
+      <Divider />
+      <List>
+        <ListItem button key={'About'} onClick={() => props.handleMenuClick('About')}>
+          <ListItemIcon><InfoIcon /></ListItemIcon>
+          <ListItemText primary={'About'} />
+        </ListItem>
+        <ListItem button key={'Close'}>
+          <ListItemIcon><ExitToAppIcon /></ListItemIcon>
+          <ListItemText primary={'Close'} />
+        </ListItem>
+      </List>
+    </div>
+  );
+
+  return (
+    <div>
+        <Drawer open={props.drawer} onClose={toggleDrawer(false)}>
+          {list()}
+        </Drawer>
+    </div>
+  );
+}
 
 export default class GuiMenu extends React.Component {
   constructor(props) {
